Made parent path var more acceptible

+++ install.sh
@@ -1,14 +1,15 @@
 #!/bin/bash
 ## Author:	Owen Cocjin
-## Version:	1.2
+## Version:	1.3
 ## Date:    07/12/20
 ## Title:   install.sh
 ## Description: Installs BashTools
 ## Notes:
 ##    - Must run as root
-##    - Changed bashtools path env var
-##    - Fixed issue with home path being incorrect when run by root
-
+##    - Prevents error when multiple users run install script
+##    - Made home path variable more accessible
+#Edit this path to whatever home directory
+parentPath='/usr/local/bin'
 #Exit if not running as root
 if [[ $(id -u) != '0' ]]; then
 	echo "[|X] Run as root (would probably have failed otherwise)!"
@@ -19,12 +20,14 @@ fi
 
 #Create temp var for bashtools home directory
 ### CHANGE ME IF YOU WANT ###
-temp_path='/usr/local/bin/bashtools'
+temp_path="${parentPath}/bashtools"
 echo "[|X] Deciding home for bashtools... ${temp_path}"
 
-#Sym link Tools to path
-ln -s ${PWD}/Tools ${temp_path}
-echo -e "[|X] Creating symlink from\n\t-> ${temp_path} to ${PWD}/Tools"
+#Sym link Tools to path if one doesn't already exist
+if [[ ! -f ${temp_path} ]]; then
+	ln -s ${PWD}/Tools ${temp_path}
+	echo -e "[|X] Creating symlink from\n\t-> ${temp_path} to ${PWD}/Tools"
+fi
 
 #Add lines to .bashrc
 echo -e "\nexport BASHTOOLS_PATH=${temp_path}
